create a controller for appointments
